ref: major codebase refactoring
