+++ src/configuration/state.rs
@@ -1,10 +1,14 @@
+//! Shared readiness flags consulted by [`crate::server::create_readiness_interceptor`]
+//! to reject gRPC requests before Aquila's upstream dependencies are usable,
+//! rather than accepting them and failing partway through a handler.
+
 use std::sync::Arc;
 use std::sync::atomic::{AtomicBool, Ordering};
 
 /// Tracks readiness of each external service.
 #[derive(Clone)]
 pub struct AppReadiness {
-    // Readiness state of Sagittarius service
+    /// Whether the Sagittarius gRPC channel has completed its initial connect.
     pub sagittarius_ready: Arc<AtomicBool>,
 }
