+++ src/sagittarius/flow_service_client_impl/flow_store.rs
@@ -1,0 +1,100 @@
+//! KV-store operations backing the flows Aquila keeps synchronized with
+//! Sagittarius: delete-by-id, wholesale replace, and single-flow upsert.
+//! Every flow is keyed by [`get_flow_identifier`], so lookups by id require
+//! a full key scan rather than a direct get.
+
+use futures::{StreamExt, TryStreamExt};
+use prost::Message;
+use tucana::shared::{Flows, ValidationFlow};
+
+use crate::flow::{get_flow_identifier, key_has_flow_id};
+
+/// Deletes every stored key matching `flow_id`, returning how many were removed.
+pub(super) async fn delete_flow(store: &async_nats::jetstream::kv::Store, flow_id: i64) -> usize {
+    let mut keys = match store.keys().await {
+        Ok(keys) => keys.boxed(),
+        Err(err) => {
+            log::error!(
+                "Failed to list stored flows for deletion flow_id={} error={:?}",
+                flow_id,
+                err
+            );
+            return 0;
+        }
+    };
+
+    let mut deleted_count = 0;
+    while let Ok(Some(key)) = keys.try_next().await {
+        if !key_has_flow_id(&key, flow_id) {
+            continue;
+        }
+
+        match store.delete(&key).await {
+            Ok(_) => deleted_count += 1,
+            Err(err) => log::error!(
+                "Failed to delete stored flow flow_id={} key={} error={:?}",
+                flow_id,
+                key,
+                err
+            ),
+        }
+    }
+
+    deleted_count
+}
+
+/// Stores a single flow update under its computed key, returning that key
+/// for the caller's own logging/metrics.
+pub(super) async fn store_flow(
+    store: &async_nats::jetstream::kv::Store,
+    flow: &ValidationFlow,
+) -> (String, Result<(), async_nats::jetstream::kv::PutError>) {
+    let key = get_flow_identifier(flow);
+    let bytes = flow.encode_to_vec();
+    let result = store.put(key.clone(), bytes.into()).await.map(|_| ());
+    (key, result)
+}
+
+/// Purges every currently stored flow and replaces it with `flows`.
+/// Returns `(purged_count, stored_count)` so the caller can report metrics.
+pub(super) async fn replace_all(
+    store: &async_nats::jetstream::kv::Store,
+    flows: Flows,
+) -> (usize, usize) {
+    let mut keys = match store.keys().await {
+        Ok(keys) => keys.boxed(),
+        Err(err) => {
+            log::error!(
+                "Failed to list stored flows before replacement error={:?}",
+                err
+            );
+            return (0, 0);
+        }
+    };
+
+    let mut purged_count = 0;
+    while let Ok(Some(key)) = keys.try_next().await {
+        match store.purge(&key).await {
+            Ok(_) => purged_count += 1,
+            Err(err) => log::error!("Failed to purge stored flow key={} error={}", key, err),
+        }
+    }
+
+    let mut stored_count = 0;
+    for flow in flows.flows {
+        let (key, result) = store_flow(store, &flow).await;
+        match result {
+            Ok(()) => {
+                stored_count += 1;
+                log::debug!("Stored replacement flow key={}", key);
+            }
+            Err(err) => log::error!(
+                "Failed to store replacement flow key={} error={:?}",
+                key,
+                err
+            ),
+        }
+    }
+
+    (purged_count, stored_count)
+}
