+++ src/server/mod.rs
@@ -1,4 +1,8 @@
-mod action_transfer_service_server_impl;
+//! Aquila's gRPC surface: the services actions and runtimes connect to.
+//! [`dynamic_server`] and [`static_server`] assemble the actual service set
+//! for each run mode from the individual `*_service_server_impl` modules.
+
+mod action_transfer;
 mod interceptor;
 mod module_service_server_impl;
 mod runtime_execution_service_server_impl;
